revert(bundle): remove parallel transaction bundle execution
Reverts the capability added by 1941ba7, f9eaa60, 4c868e7 and documented in
1707266. It never earned its keep on the workloads we run, and measurement on
the perf box showed it is actively harmful when switched on.

Measured, 1000-bundle Synthea import at 20 VUs against Postgres capped at 8
CPUs and a 30G cgroup (NVMe storage, identical data each run):

  K=1 (capability off)             3.540 it/s    0 failures
  K=4, processingDefault=sequential 3.608 it/s   0 failures   (+1.9%, noise)
  K=2, processingDefault=parallel  3.410 it/s    0 failures   (-3.7%)
  K=4, processingDefault=parallel  --            391/1000 failed, Postgres OOM

At K=4 with parallel actually engaged, 20 concurrent bundles x 4 shards put 80
backends against shared_buffers=10GB inside the container's memory limit; the
kernel OOM-killed the checkpointer and Postgres went into crash recovery:

  Memory cgroup out of memory: Killed process (postgres) shmem-rss:7738400kB
  LOG: checkpointer process was terminated by signal 9: Killed

Note the startup guard only checked connections
(expected_concurrent_bundles * transactionConcurrency <= pool max_conns, which
80 <= 128 satisfied). Memory was the binding constraint and nothing checked it.

The shape of the workload is why there is no upside: sharding within a bundle
only helps when client concurrency sits below the available cores, and 20
concurrent bundles already oversubscribe 8 CPUs ~2.5x. K=2 being slower than
K=1 is the coordination overhead with no headroom to recover it.

Removes the executor, the BundleTuning store option, both config tunables
(bundle.transactionConcurrency / bundle.transactionProcessingDefault and their
env vars), the x-bundle-processing-logic header plumbing, and the tuning guide
section. Unrecognised values of that header were already ignored rather than
rejected, so dropping it changes no client-visible behaviour.

verbOrder moves back into bundle.go: it lived in the parallel file but is used
by the serial path. internal/index/extractor_race_test.go is kept — it proves
the extractor is goroutine-safe, which stands on its own.
