+++ internal/handler/bundle_test.go
@@ -17,11 +17,8 @@
 package handler_test
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/wso2/fhir-server/internal/store"
@@ -131,58 +128,5 @@ func TestBundle_TrailingSlashAlsoRoutes(t *testing.T) {
 	})
 	if resp.Code != http.StatusOK || !called {
 		t.Fatalf("trailing-slash POST did not route to bundle handler: status=%d called=%v", resp.Code, called)
-	}
-}
-
-// TestBundle_ProcessingLogicHeader verifies x-bundle-processing-logic plumbing:
-// a recognised value (case-insensitive) reaches the store via the context, an
-// unrecognised value or absent header leaves the context unset, so the store's
-// configured default applies.
-func TestBundle_ProcessingLogicHeader(t *testing.T) {
-	cases := []struct {
-		name   string
-		header string
-		want   string // expected store.BundleProcessingFrom(ctx)
-	}{
-		{"parallel", "parallel", store.BundleProcessingParallel},
-		{"case-insensitive", "PaRaLLeL", store.BundleProcessingParallel},
-		{"sequential", "Sequential", store.BundleProcessingSequential},
-		{"unrecognised ignored", "bogus", ""},
-		{"absent", "", ""},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			var got string
-			ms := &mockStore{
-				executeBundleFn: func(ctx context.Context, _, _ string, _ []store.BundleEntryRequest) ([]store.BundleEntryResult, error) {
-					got = store.BundleProcessingFrom(ctx)
-					return nil, nil
-				},
-			}
-			h := newRouter(ms)
-
-			body, _ := json.Marshal(map[string]any{
-				"resourceType": "Bundle",
-				"type":         "transaction",
-				"entry": []any{map[string]any{
-					"resource": map[string]any{"resourceType": "Patient"},
-					"request":  map[string]any{"method": "POST", "url": "Patient"},
-				}},
-			})
-			req := httptest.NewRequest(http.MethodPost, "/fhir/r4", bytes.NewReader(body))
-			req.Header.Set("Content-Type", "application/fhir+json")
-			if tc.header != "" {
-				req.Header.Set("x-bundle-processing-logic", tc.header)
-			}
-			w := httptest.NewRecorder()
-			h.ServeHTTP(w, req)
-
-			if w.Code != http.StatusOK {
-				t.Fatalf("status = %d, want 200; body=%s", w.Code, w.Body.String())
-			}
-			if got != tc.want {
-				t.Errorf("store saw processing logic %q, want %q", got, tc.want)
-			}
-		})
 	}
 }
